keep raw pixels through dp-sharded ViT path

+++ vllm/model_executor/models/qwen2_5_vl.py
@@ -1445,7 +1445,7 @@ class Qwen2_5_VLForConditionalGeneration(
             if self.use_data_parallel:
                 return run_dp_sharded_mrope_vision_model(
                     self.visual,
-                    pixel_values.type(self.visual.dtype),
+                    pixel_values,
                     grid_thw_list,
                     rope_type="rope_3d",
                 )
@@ -1505,7 +1505,7 @@ class Qwen2_5_VLForConditionalGeneration(
             if self.use_data_parallel:
                 return run_dp_sharded_mrope_vision_model(
                     self.visual,
-                    pixel_values_videos.type(self.visual.dtype),
+                    pixel_values_videos,
                     grid_thw_list,
                     rope_type="rope_3d",
                 )
